# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "summary": "Root",
        "operationId": "root__get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/groups/{group_id}/students": {
      "get": {
        "summary": "Read Group",
        "operationId": "read_group_groups__group_id__students_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyGroup"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/students/{student_id}/activities": {
      "get": {
        "summary": "Get Activities",
        "operationId": "get_activities_students__student_id__activities_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/ActivityType"
            }
          },
          {
            "name": "min_last_changed_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Min Last Changed Date"
            }
          },
          {
            "name": "max_last_changed_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Max Last Changed Date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Activity"
                  },
                  "title": "Response Get Activities Students  Student Id  Activities Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Post Activity",
        "operationId": "post_activity_students__student_id__activities_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "student_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Student Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActivityCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Activity"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/groups/{group_id}/stats": {
      "get": {
        "summary": "Get Stats",
        "operationId": "get_stats_groups__group_id__stats_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "group_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GroupStats"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Activity": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "activity_type": {
            "$ref": "#/components/schemas/ActivityType"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "last_changed": {
            "type": "string",
            "format": "date",
            "title": "Last Changed"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "student_id",
          "activity_type",
          "last_changed",
          "created_at"
        ],
        "title": "Activity"
      },
      "ActivityCreate": {
        "properties": {
          "activity_type": {
            "$ref": "#/components/schemas/ActivityType"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "last_changed": {
            "type": "string",
            "format": "date",
            "title": "Last Changed"
          }
        },
        "type": "object",
        "required": [
          "activity_type",
          "last_changed"
        ],
        "title": "ActivityCreate"
      },
      "ActivityType": {
        "type": "string",
        "enum": [
          "lesson_completed",
          "quiz_attempted",
          "note_added"
        ],
        "title": "ActivityType"
      },
      "GroupStats": {
        "properties": {
          "group_id": {
            "type": "integer",
            "title": "Group Id"
          },
          "total_activities": {
            "type": "integer",
            "title": "Total Activities"
          },
          "average_quiz_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Average Quiz Score"
          }
        },
        "type": "object",
        "required": [
          "group_id",
          "total_activities"
        ],
        "title": "GroupStats"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "StudentBase": {
        "properties": {
          "student_id": {
            "type": "string",
            "title": "Student Id"
          },
          "fname": {
            "type": "string",
            "title": "Fname"
          },
          "lname": {
            "type": "string",
            "title": "Lname"
          },
          "last_changed": {
            "type": "string",
            "format": "date",
            "title": "Last Changed"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "fname",
          "lname",
          "last_changed"
        ],
        "title": "StudentBase"
      },
      "StudyGroup": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "group_name": {
            "type": "string",
            "title": "Group Name"
          },
          "last_changed": {
            "type": "string",
            "format": "date",
            "title": "Last Changed"
          },
          "students": {
            "items": {
              "$ref": "#/components/schemas/StudentBase"
            },
            "type": "array",
            "title": "Students",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "group_name",
          "last_changed"
        ],
        "title": "StudyGroup"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}